refactor(sozluk): collapse Term wire shape onto a single term-fields.ts map (#1544)
Term's wire field set was restated across three hand-synced sites: the
`TermView` field literal (views.ts), the `TermFields` interface + `toTerm`
shaper (shapers.ts), and the `TermSummaryRow`/`toTermSummaryRow` row mapper
(term-summary.ts). A field add/rename meant editing all three in lockstep.

Introduce `sozluk/term-fields.ts` as the single column→field map, mirroring
the proven `definition-fields.ts` template: a `termViewFields` literal pinned
`as const satisfies Record<keyof TermSummaryRow, true>`, with `TermSummaryRow`
and `toTermSummaryRow` derived from one `intrinsicFields` declaration whose
readers absorb the `definitionCount`→`count`, `lastEditAt`→`lastEdit`, and
slug-as-`id` divergence. `TermView` consumes `...termViewFields`, the standalone
`TermFields` interface is gone (`toTerm` takes the map-derived `TermSummaryRow`),
and `term-summary.ts` is folded into the new single source.

Behavior-preserving: the `Term` wire shape is byte-identical — the fate-generated
client is unchanged. Dropping a field from the map (or adding one to the row type
without listing it) is now a compile error; wire-shape drift is structurally
unrepresentable, not test-detected.

Part of the fate-wire collapse epic #1332.
